pyproject.toml
[project]
name = "hack-nu"
version = "0.1.0"
description = "Evaluates solutions on a given dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "qdrant-client>=1.17.0",
    "rank-bm25>=0.2.2",
    "numpy>=2.4.2",
    "openai>=2.21.0",
]
